% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assurance_process_data_move.R
\name{assurance_process_data_move}
\alias{assurance_process_data_move}
\title{Assurance process data move}
\usage{
assurance_process_data_move(
  period.from = NULL,
  period.to = NULL,
  onedrive.path = NULL,
  conf.loc = NULL,
  tp_sites = NULL,
  folder = "Data Assurance",
  combine = FALSE,
  prem.col = NULL,
  meter.col = NULL
)
}
\arguments{
\item{period.from}{A date. Files that were created after this date will be processed. Defaults to the date exactly a month before the current date.}

\item{period.to}{A date. Files that were created before this date will be processed. Defaults to the next day from the current date, to pick up all of the files created on the current day.}

\item{onedrive.path}{Path to where all of the OneDrive folders are synced. If NULL, function attempts to guess this.}

\item{conf.loc}{Location of the configuration file that keeps the Blob storage access keys. If NULL, a pop-up window appears, which can be used to locate the file}

\item{tp_sites}{A list of all the Trading Party sites from where to look for files. By default all of the Trading Party sites in the Onedrive folder will be used}

\item{folder}{Folder in the Trading Party site, where the files are kept}

\item{combine}{A boolean, indicating whether the raw files should be combined into one file}

\item{prem.col}{The column names of the Premises file template}

\item{meter.col}{The column names of the meter file template}
}
\value{
Returns a table which contains information about all of the files processed
}
\description{
This function moves data from Trading Party Sharepoint sites
to a MOSL blob storage. It also tracks all of the changes and
validates the received data. Validated files are then sent back
to the Trading Party's Sharepoint site
}
