% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_chart.R
\name{bubble_chart}
\alias{bubble_chart}
\title{Plot Bubble Graph}
\usage{
bubble_chart(
  df = NULL,
  vertical.sep = 0.15,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  x_lab = NULL,
  y_lab = NULL
)
}
\arguments{
\item{df}{dataframe}

\item{vertical.sep}{numerical}

\item{title}{character}

\item{subtitle}{character}

\item{caption}{character}

\item{my.dir}{character}

\item{x}{character}

\item{y}{character}
}
\value{
A bubble graph for MPOP reporting
}
\description{
This function creates a bubble graph, based
on either long unread data or vacancy data,
where on the x-axis is the current vacancy
or longunread level and on the y-axis is the
change in the longunread or vacancy level
compared to the level 12 months ago
}
