% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_assurance_process.R
\name{check_for_templates}
\alias{check_for_templates}
\title{Check For Templates}
\usage{
check_for_templates(template.folders, excluded = "MOSL-Reviewed")
}
\arguments{
\item{template.folders}{List of folders that could contain templates}

\item{excluded}{Pattern for files to be excluded from output}
}
\description{
Check For Templates
}
