% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_monthly_tracker_mps.R
\name{create_monthly_tracker}
\alias{create_monthly_tracker}
\title{Create Monthly MPS Tracker (Pre Analysis and PfM Commentary)}
\usage{
create_monthly_tracker(
  period = Sys.Date() \%m-\% months(1),
  mps_list = c("MPS 1", "MPS 2", "MPS 3", "MPS 4", "MPS 5", "MPS 7", "MPS 12", "MPS 16",
    "MPS 17", "MPS 18"),
  my.dir = getwd(),
  rda.outputs = paste0(my.dir, "/data/rdata"),
  dir.mps.tracking = paste0(my.dir, "/data/tracking/mps"),
  period.only = TRUE
)
}
\arguments{
\item{period}{date}

\item{mps_list}{character}

\item{my.dir}{character}

\item{rda.outputs}{character}

\item{dir.mps.tracking}{character}

\item{period.only}{logical}
}
\value{

}
\description{
Compares the performance across MPS for each Trading Party
by MPS and compares IPRP performance versus planned milestones.
This is done prior to analysis and Pfm commentary being added.
}
