% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tracker.R
\name{create_tracker}
\alias{create_tracker}
\title{Creates Monthly Tracker (Pre Analysis and PfM Commentary)}
\usage{
create_tracker(
  period = Sys.Date() \%m-\% months(1),
  my.dir = getwd(),
  standard.data,
  iprp.plans,
  tracking.sheet,
  standard,
  rda.outputs = paste0(my.dir, "/data/rdata"),
  dir.tracking,
  period.only = TRUE,
  save.output = TRUE,
  keep.vars = FALSE,
  iprp.list,
  filter.category = c("IPRP: on-track", "Normal monitoring")
)
}
\arguments{
\item{period}{date}

\item{my.dir}{character}

\item{standard.data}{dataframe}

\item{iprp.plans}{dataframe}

\item{tracking.sheet}{dataframe}

\item{standard}{character}

\item{rda.outputs}{character}

\item{dir.tracking}{character}

\item{period.only}{logical}

\item{save.output}{logical}

\item{keep.vars}{logical}

\item{iprp.list}{character}

\item{filter.category}{character}
}
\value{

}
\description{
Compares the performance across a given standard for each Trading Party
by Standard and compares IPRP performance versus planned milestones.
This is done prior to analysis and Pfm commentary being added.
}
