% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_assurance_process.R
\name{data_assurance_process}
\alias{data_assurance_process}
\title{Data Assurance Process}
\usage{
data_assurance_process(
  onedrive.folder = "Data Assurance/New",
  conf.loc = NULL,
  config.name = "data-assurance",
  upload.blob = TRUE,
  clean.sharepoint = TRUE,
  upload.log = TRUE,
  ...
)
}
\arguments{
\item{onedrive.folder}{Folder or directory within Sharepoint Org site that should contain the new templates to be processed}

\item{conf.loc}{The path to a configuration file containing Azure storage connection information and keys}

\item{upload.blob}{Logical specifying whether to upload the new, processed and rejected templates to blob storage}

\item{clean.sharepoint}{Logical specifying whether to clean-up Sharepoint by moving the processed file}

\item{upload.log}{Logical specifying whether to upload the log file to blob storage}

\item{...}{Optional arguments for underlying functions}
}
\description{
GET TEMPLATE PATHS FROM ONEDRIVE
Get possible synced OneDrive locations for templates in Data Assurance folder
Check folder exists, if not show warning and skip
Check if relevant files in folder, if not show message and skip
Check if CSV
}
\details{
UPLOAD TEMPLATES TO BLOB NEW
Use config file to store container endpoint and SAS key
If relevant files present:
Clean columns to upper_camel
Upload to blob container under New, saving file as "ORG-ID_FileCreatedDate_OriginalFileName"

VERIFY TEMPLATES ARE VALID AND VALIDATE COLUMNS
Check columns in file match expected columns for either Premises or Meters
Verify columns are valid:
Measure
Action
Explanation
Create Status columns based on validations:
ValidationStatus: Pass / Fail
Reason: NA / Validation failure reasons

UPLOAD PROCESSED FILES TO BLOB PROCESSED OR REJECTED
Include relevant Blob metadata: ORG-ID, FileCreatedDate, OriginalFileName
Copy file to OneDrive
Save file as "ORG-ID_FileCreatedDate_OriginalFileName_MOSL-Reviewed_ReviewDate.csv"
For valid rows:
Copy to Blob Processed
Save file as "ORG-ID_FileCreatedDate_OriginalFileName.csv"
For invalid rows:
Copy to Blob Rejected
Save file as "ORG-ID_FileCreatedDate_OriginalFileName.csv"
}
