% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_copy.R
\name{data_copy}
\alias{data_copy}
\title{Copy data from one database to another}
\usage{
data_copy(
  my.dir = getwd(),
  conf.loc = NULL,
  conf.from = "datamgmt3",
  conf.to = "sandpit",
  query = NULL,
  saveDB = TRUE,
  output.table = "PERF_OPSData",
  overwrite = TRUE,
  append = FALSE,
  save.csv = FALSE,
  csv.file = "/data/inputs/OPS_data.csv"
)
}
\arguments{
\item{my.dir}{working directory}

\item{query}{SQL query that gets the source data}

\item{saveDB}{should the data be saved to a new database}

\item{output.table}{name of the output table in a database}

\item{overwrite}{should the existing table be overwritten}

\item{append}{should the existing table be appended to}

\item{save.csv}{should a csv file be saved}

\item{csv.file}{name of the csv file}

\item{server.from}{server name for the source data}

\item{server.to}{server name to where data is copied to}

\item{db.to}{database name to where data is copied to}
}
\description{
data_copy() is designed to copy data from one database to another.
By default, it copies the OPS data from data-mgmt3 to data-mgmt.
Can also save the table as .csv file
}
