% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragon_drop.R
\name{dragon_drop}
\alias{dragon_drop}
\title{Drag and drop files from one location to another based on folder name to a specified directory}
\usage{
dragon_drop(
  dir.to = getwd(),
  dir.from = utils::choose.dir(caption =
    "Select the directory that folders  should be copied from."),
  folder.name = format(Sys.Date(), format = "\%Y-\%m"),
  folder.ext = NULL,
  sub.folder = NULL,
  file.type = NULL,
  tp.list = "to",
  new.folder = TRUE
)
}
\arguments{
\item{dir.to}{character}

\item{dir.from}{character}

\item{folder.name}{character}

\item{folder.ext}{character}

\item{sub.folder}{character}

\item{file.type}{factor}

\item{tp.list}{factor}

\item{new.folder}{character}
}
\value{

}
\description{
Drag and drop files from one location to another based on folder name to a specified directory
}
