% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config.R
\name{get_config}
\alias{get_config}
\title{Get Config File}
\usage{
get_config(conf.loc = NULL, config.name = "default")
}
\arguments{
\item{conf.loc}{Path to config file that can be specified upfront.}

\item{config.name}{Name of the section of the config file that will be used.}
}
\description{
This function wraps around config::get and adds additional logic to
locate and return a relevant config file and sets the needed
environment variables.
}
