% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_assurance_process.R
\name{get_onedrive_path}
\alias{get_onedrive_path}
\title{Get OneDrive Path}
\usage{
get_onedrive_path(
  manual.select = FALSE,
  root.folder = "Market Operator Services Limited"
)
}
\arguments{
\item{manual.select}{Option for whether to select the folder manually.}

\item{root.folder}{String specifying the name of the root folder (e.g. company name) for OneDrive.}
}
\description{
Get OneDrive Path
}
