% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_assurance_process.R
\name{get_template_paths_from_onedrive}
\alias{get_template_paths_from_onedrive}
\title{Get Template Paths From OneDrive}
\usage{
get_template_paths_from_onedrive(root.folder, onedrive.folder)
}
\arguments{
\item{root.folder}{String specifying the name of the root folder (e.g. company name) for OneDrive.}

\item{onedrive.folder}{Folder or directory within Sharepoint Org site that should contain the new templates to be processed}
}
\description{
Get Template Paths From OneDrive
}
