% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSL_colours.R
\name{MOSL_colours}
\alias{MOSL_colours}
\title{MOSL_colours}
\usage{
MOSL_colours(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
A function containing all of MOSL brand colours
}
\details{
The available colours are: 'grey', 'light blue', 'dark blue', green', 'orange', 'yellow' and 'purple'
}
