% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSL_palettes.R
\name{MOSL_palettes}
\alias{MOSL_palettes}
\title{MOSL_palettes}
\usage{
MOSL_palettes(palette = "all", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Available palettes are:\cr
1) 'core', (includes colours 'grey', 'light blue' and 'dark blue')\cr
2) 'secondary' (includes colours 'green', 'orange', 'yellow' and 'purple') \cr
3) 'all' (includes all 7 previously mentioned colours)}

\item{...}{}
}
\value{

}
\description{
A function containing three palettes with either MOSL's core colours, secondary colours or all colours.
}
\examples{

}
