% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPOP_data_prep.R
\name{MPOP_data_prep}
\alias{MPOP_data_prep}
\title{Prepare MPOP report data}
\usage{
MPOP_data_prep(
  df = NULL,
  df.tp = NULL,
  output.table = c("vacancy", "longunread"),
  by = "Retailer"
)
}
\arguments{
\item{df}{dataframe}

\item{df.tp}{character}

\item{output.table}{character}

\item{by}{character}
}
\value{

}
\description{
Prepares data for bubble charts and/or Trading
Party ranking tables in MPOP reporting
}
\examples{

}
