% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps_aggregate_perf.R
\name{mps_aggregate_perf}
\alias{mps_aggregate_perf}
\title{Summarise Trading Parties by MPS}
\usage{
mps_aggregate_perf(
  my.dir = getwd(),
  df = readRDS(paste0(my.dir, "/data/rdata/mps_data_clean.Rda")),
  tp.details = utils::read.csv(paste0(my.dir, "/data/inputs/tp_details.csv")),
  spid.counts = utils::read.csv(paste0(my.dir, "/data/inputs/spid_counts.csv"))
)
}
\arguments{
\item{my.dir}{character}

\item{df}{dataframe}

\item{tp.details}{dataframe}
}
\description{
Summarise Trading Parties by MPS
}
