% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps_data_melt.R
\name{mps_melt_data}
\alias{mps_melt_data}
\title{Melt MPS Data}
\usage{
mps_melt_data(
  my.dir = getwd(),
  mps.data = readRDS(paste0(my.dir, "/data/rdata/mps_data_clean.Rda")),
  filter.date.from = "2018-04-01",
  filter.date.to = Sys.Date(),
  iprp.plans = utils::read.csv(paste0(my.dir, "/data/inputs/IPRP_plans_mps.csv")),
  save.output = TRUE,
  save.dir = paste0(my.dir, "/data/rdata/mps_data_melt.Rda")
)
}
\arguments{
\item{my.dir}{character}

\item{mps.data}{file}

\item{filter.date.from}{date}

\item{filter.date.to}{date}

\item{iprp.plans}{file}

\item{save.output}{logical}

\item{save.dir}{character}
}
\value{

}
\description{
This function pivots the the MPS data
so that performance can be plotted in a
graph.
}
