% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps_performance.R
\name{mps_performance}
\alias{mps_performance}
\title{MPS Performance}
\usage{
mps_performance(
  period.lag = 1,
  mps_list = c("MPS 1", "MPS 2", "MPS 3", "MPS 4", "MPS 7", "MPS 12", "MPS 16",
    "MPS 17", "MPS 18"),
  dir = getwd(),
  dir.mps.data = paste0(dir, "/data/rdata/mps_data_clean.Rda"),
  dir.iprp.plans = paste0(dir, "/data/inputs/IPRP_plans_mps.csv"),
  dir.watch.list = paste0(dir, "/data/inputs/tracking_mps.csv"),
  csv.outputs = paste0(dir, "/data/outputs"),
  rda.outputs = paste0(dir, "/data/rdata"),
  dir.mps.tracking = paste0(dir, "/data/tracking/mps")
)
}
\arguments{
\item{period.lag}{numeric}

\item{mps_list}{character}

\item{dir}{character}

\item{dir.mps.data}{character}

\item{dir.iprp.plans}{character}

\item{dir.watch.list}{character}

\item{csv.outputs}{character}

\item{rda.outputs}{character}

\item{dir.mps.tracking}{character}
}
\value{

}
\description{
MPS Performance
}
