% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps_plot_perf_graph.R
\name{mps_plot_perf_graph}
\alias{mps_plot_perf_graph}
\title{Plot Performance Graph (MPS)}
\usage{
mps_plot_perf_graph(
  df,
  load.data = FALSE,
  my.dir = NULL,
  include.iprp = FALSE,
  trading.party,
  standard,
  period.start = "2018-04-01",
  period.end = Sys.Date(),
  action.points = FALSE,
  sec.axis.name = "Performance",
  fill.label = "Task Volume",
  fill.manual = "azure3",
  y.lab = "Tasks",
  x.lab = "",
  graph.title = NULL,
  sub.title = NULL
)
}
\arguments{
\item{df}{dataframe}

\item{load.data}{logical}

\item{my.dir}{character}

\item{include.iprp}{logical}

\item{trading.party}{character}

\item{standard}{character}

\item{period.start}{date}

\item{period.end}{date}

\item{action.points}{logical}

\item{sec.axis.name}{character}

\item{fill.label}{character}

\item{fill.manual}{character}

\item{y.lab}{character}

\item{x.lab}{chracter}

\item{graph.title}{character}

\item{sub.title}{character}
}
\value{

}
\description{
This function produces a grasph with two #' independent
axis. The left-handside corresponds to a single variable
plotted as a bar, and the right-handside corresponds to
multiple lines.
}
