% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps_sum_TradingParties.R
\name{mps_sum_TradingParties}
\alias{mps_sum_TradingParties}
\title{Summarise Trading Parties by MPS}
\usage{
mps_sum_TradingParties(
  my.dir = getwd(),
  df = readRDS(paste0(my.dir, "/data/rdata/mps_data_clean.Rda")),
  tp.details = utils::read.csv(paste0(my.dir, "/data/inputs/tp_details.csv"))
)
}
\arguments{
\item{my.dir}{character}

\item{mps_data_clean}{character}
}
\value{

}
\description{
Summarise Trading Parties by MPS
}
