% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_aggregate_perf.R
\name{ops_aggregate_perf}
\alias{ops_aggregate_perf}
\title{Aggregates Trading Party OPS Performance}
\usage{
ops_aggregate_perf(
  my.dir = getwd(),
  df = readRDS(paste0(my.dir, "/data/rdata/ops_data_clean.Rda")),
  tp.details = utils::read.csv(paste0(my.dir, "/data/inputs/tp_details.csv")),
  spid.counts = utils::read.csv(paste0(my.dir, "/data/inputs/spid_counts.csv"))
)
}
\arguments{
\item{my.dir}{character}

\item{df}{dataframe}

\item{tp.details}{dataframe}
}
\description{
Aggregates Trading Party OPS Performance
}
