% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_create_tracker.R
\name{ops_create_tracker}
\alias{ops_create_tracker}
\title{Create Monthly OPS Tracker (Pre Analysis and PfM Commentary)}
\usage{
ops_create_tracker(
  period = Sys.Date() \%m-\% months(1),
  my.dir = getwd(),
  rda.outputs = paste0(my.dir, "/data/rdata"),
  dir.ops.tracking = paste0(my.dir, "/data/tracking/ops"),
  period.only = TRUE,
  save.output = TRUE,
  keep.vars = FALSE,
  iprp.list = NULL,
  filter.category = c("IPRP: on-track", "Normal monitoring",
    "Performance flag: 6 month")
)
}
\arguments{
\item{period}{date}

\item{my.dir}{character}

\item{rda.outputs}{character}

\item{dir.ops.tracking}{character}

\item{period.only}{logical}

\item{save.output}{logical}

\item{keep.vars}{logical}

\item{iprp.list}{character}

\item{filter.category}{character}
}
\value{

}
\description{
Compares the performance across OPS for each Trading Party
by OPS and compares IPRP performance versus planned milestones.
This is done prior to analysis and Pfm commentary being added.
}
