% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_performance.R
\name{ops_performance}
\alias{ops_performance}
\title{OPS Performance}
\usage{
ops_performance(
  period.lag = 1,
  ops.list = c("OPS B5a", "OPS C1a"),
  dir = getwd(),
  dir.ops.data = paste0(dir, "/data/rdata/ops_data_clean.Rda"),
  dir.iprp.plans = paste0(dir, "/data/inputs/IPRP_plans_ops.csv"),
  dir.watch.list = paste0(dir, "/data/inputs/tracking_ops.csv"),
  csv.outputs = paste0(dir, "/data/outputs"),
  rda.outputs = paste0(dir, "/data/rdata"),
  dir.ops.tracking = paste0(dir, "/data/tracking/ops")
)
}
\arguments{
\item{period.lag}{numeric}

\item{ops.list}{character}

\item{dir}{character}

\item{dir.ops.data}{character}

\item{dir.iprp.plans}{character}

\item{dir.watch.list}{character}

\item{csv.outputs}{character}

\item{rda.outputs}{character}

\item{dir.ops.tracking}{character}
}
\value{

}
\description{
Compares the performance across OPS for a given period
for each Trading Party; and compares IPRP performance
versus planned milestones.
}
