% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_process_tracker.R
\name{ops_process_tracker}
\alias{ops_process_tracker}
\title{Process Monthly OPS Tracker (Post Analysis and PfM Commentary)}
\usage{
ops_process_tracker(
  my.dir = getwd(),
  save.output = TRUE,
  save.dir.rds = paste0(my.dir, "/data/rdata/perf_status_ops.Rda"),
  save.dir.csv = paste0(my.dir, "/data/outputs/perf_status_ops.csv"),
  keep.vars = TRUE,
  period.create = Sys.Date() \%m-\% months(1),
  DataBase = TRUE
)
}
\arguments{
\item{my.dir}{character}

\item{save.output}{logical}

\item{save.dir.rds}{character}

\item{save.dir.csv}{character}

\item{keep.vars}{logical}

\item{period.create}{date}
}
\value{

}
\description{
This function processes the OPS performance data for each
Trading Party by OPS after the analysis and PfM Commentary
has been added. This will then be used to  render the PfM
and Monthly Performance Reports created each month.
}
