% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_perf_graphs_all.R
\name{plot_perf_graphs_all}
\alias{plot_perf_graphs_all}
\title{Create Performance graphs}
\usage{
plot_perf_graphs_all(
  my.dir = getwd(),
  df.mps = NULL,
  df.ops = NULL,
  load.data = TRUE,
  output.dir = paste0(my.dir, "/graphs"),
  run.parallel = FALSE
)
}
\arguments{
\item{my.dir}{character}

\item{load.data}{boolean}

\item{output.dir}{character}

\item{run.parallel}{boolean}

\item{df}{dataframe}
}
\value{

}
\description{
This function loads the MPS and OPS data
and creates all MPS and OPS performance graphs
}
\examples{

}
