% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_mps_render_monthly_pfm_report.R
\name{prep_mps_render_monthly_pfm_report}
\alias{prep_mps_render_monthly_pfm_report}
\title{Prepare MPS Data for Monthly PfM Report}
\usage{
prep_mps_render_monthly_pfm_report(
  my.dir = getwd(),
  filter.date.from = "2018-04-01",
  filter.date.to = Sys.Date()
)
}
\arguments{
\item{my.dir}{character}

\item{filter.date.from}{date}

\item{filter.date.to}{date}
}
\value{

}
\description{
This function prepares the MPS data that will be used
to render the PfM Report that is created each month.
}
