% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_render_monthly_perf_rep.R
\name{prep_render_monthly_perf_rep}
\alias{prep_render_monthly_perf_rep}
\title{Data Preperation for Monthly Performance Report}
\usage{
prep_render_monthly_perf_rep(
  dir = getwd(),
  period.lag = 1,
  dir.mps.data = paste0(dir, "/data/rdata/mps_data_clean.Rda"),
  dir.mps.sum = paste0(dir, "/data/rdata/mps_data_clean.Rda"),
  dir.iprp.comp = paste0(dir, "/data/rdata/IPRP_plan_comparison_mps.Rda"),
  tracking.mps = paste0(dir, "/data/inputs/tracking_mps.csv"),
  iprp.plans = paste0(dir, "/data/inputs/IPRP_plans_mps.csv"),
  filter.date = "2018-04-01",
  rda.outputs = paste0(dir, "/data/rdata"),
  csv.outputs = paste0(dir, "/data/outputs"),
  dir.output = paste0(dir, "/MonthlyPerfReport")
)
}
\arguments{
\item{dir}{character}

\item{period.lag}{numeric}

\item{dir.mps.data}{character}

\item{dir.mps.sum}{character}

\item{dir.iprp.comp}{character}

\item{tracking.mps}{character}

\item{iprp.plans}{character}

\item{filter.date}{date}

\item{rda.outputs}{character}

\item{csv.outputs}{character}

\item{dir.output}{character}
}
\value{

}
\description{
This function prepares the background data for rendering
the Monthly Performance Report.
}
