% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_assurance_process.R
\name{process_and_upload_templates_to_blob}
\alias{process_and_upload_templates_to_blob}
\title{Upload Assurance Templates to Blob}
\usage{
process_and_upload_templates_to_blob(
  template.path,
  cont,
  org.site.regex = "(-R)|(-W)|(-TEST)",
  org.id.regex = "[A-Z]+(-R|-W|-TEST)",
  dowload.template.func = NULL,
  dowload.template.args = NULL,
  process.template.func = "process_template_default",
  clean.sharepoint,
  upload.blob
)
}
\arguments{
\item{template.path}{Path of the file, determined in earlier step}

\item{cont}{Connection for Azure blob storage}

\item{org.site.regex}{Regular expression for identifying the name of the org site from the template path}

\item{org.id.regex}{Regular expression for identifying the org id from the template path}

\item{dowload.template.func}{User can specify a function for downloading a template from Sharepoint to e.g. change how certain columns are modified}

\item{dowload.template.args}{Optional arguments for user specified function for downloading template function}

\item{process.template.func}{User can specify a function for processing the template, but must take only one argument (i.e. template file) and return only a processed template}

\item{clean.sharepoint}{Logical specifying whether to clean Sharepoint by deleting files etc.}

\item{upload.blob}{Logical specifying whether to upload files to blob storage}
}
\description{
Upload raw templates to Blob New and then processes templates before
uploading to Blob processed and OneDrive.
}
