% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_monthly_tracker_mps.R
\name{process_monthly_tracker_mps}
\alias{process_monthly_tracker_mps}
\title{Process Monthly MPS Tracker (Post Analysis and PfM Commentary)}
\usage{
process_monthly_tracker_mps(
  my.dir = getwd(),
  filter.date.from = "2018-04-01",
  filter.date.to = Sys.Date()
)
}
\arguments{
\item{my.dir}{character}

\item{filter.date.from}{date}

\item{filter.date.to}{date}
}
\value{

}
\description{
This function processes the MPS performance data for each
Trading Party by MPS after the analysis and PfM Commentary
has been added. This will then be used to  render the PfM
and Monthly Performance Reports created each month.
}
