% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_assurance_process.R
\name{process_template_default}
\alias{process_template_default}
\title{Process Template Default}
\usage{
process_template_default(template_file, prem.col = NULL, meter.col = NULL)
}
\arguments{
\item{template_file}{The template file to be processed}

\item{prem.col}{Optional list of columns detailing expected to be in a premises template}

\item{meter.col}{Optional list of columns detailing expected to be in a meter template}
}
\description{
This function is the default function for processing a given template, and
outputs a processed file.
}
