% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tracker.R
\name{process_tracker}
\alias{process_tracker}
\title{Process Monthly Tracker (Post Analysis and PfM Commentary)}
\usage{
process_tracker(
  my.dir = getwd(),
  standard.data,
  iprp.plans,
  tracking.sheet,
  standard,
  save.output = TRUE,
  save.dir.rds = paste0(my.dir, "/data/rdata/perf_status_mps.Rda"),
  save.dir.csv = paste0(my.dir, "/data/outputs/perf_status_mps.csv"),
  keep.vars = TRUE,
  period.create = Sys.Date() \%m-\% months(1)
)
}
\arguments{
\item{my.dir}{character}

\item{standard.data}{dataframe}

\item{iprp.plans}{dataframe}

\item{tracking.sheet}{dataframe}

\item{standard}{character}

\item{save.output}{logical}

\item{save.dir.rds}{character}

\item{save.dir.csv}{character}

\item{keep.vars}{logical}

\item{period.create}{date}
}
\description{
This function processes the performance data for each Trading
Party by a given standard after the analysis and PfM Commentary
has been added. This will then be used to  render the PfM
and Monthly Performance Reports created each month.
}
