% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_lumvac_report.R
\name{render_lumvac_report}
\alias{render_lumvac_report}
\title{Render Long Unread Meter and Vacancy Report}
\usage{
render_lumvac_report(
  rmd.file = "lum-vacancy_report_main.Rmd",
  my.dir = getwd(),
  load.data = TRUE,
  dir.output = paste0(my.dir, "/LUMsVacancyReport"),
  file.name = paste0("LUMsVacancyReport-", as.character(format(Sys.Date(), "\%Y-\%m")),
    ".pdf"),
  data.period = Sys.Date() \%m-\% months(1),
  ...
)
}
\arguments{
\item{rmd.file}{character}

\item{my.dir}{character}

\item{load.data}{boolean}

\item{dir.output}{character}

\item{file.name}{character}

\item{data.period}{date}

\item{...}{}
}
\description{
This functions prepares and loads the required
datasets and renders the monthly LUMs and vacancy report
}
