% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_monthly_perf_rep.R
\name{render_monthly_perf_rep}
\alias{render_monthly_perf_rep}
\title{Render Rmarkdown Monthly Performance Report}
\usage{
render_monthly_perf_rep(
  dir = getwd(),
  dir.output = paste0(dir, "/MonthlyPerfReport"),
  file.name = paste0("MonthlyPerformanceReport-", as.character(format(Sys.Date(),
    "\%Y-\%m")), ".pdf"),
  ...
)
}
\arguments{
\item{dir}{character}

\item{dir.output}{character}

\item{file.name}{character}

\item{...}{foo}
}
\value{

}
\description{
This function renders the Monthly Performance Report
using the relevant .Rmd file.
}
