% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_monthly_pfm_report.R
\name{render_monthly_pfm_report}
\alias{render_monthly_pfm_report}
\title{Title}
\usage{
render_monthly_pfm_report(
  dir,
  rmd.file = "MonthlyPfmReport_word.Rmd",
  TRADING.PARTY,
  output.file = paste0(TRADING.PARTY, "_pfm-report_", format(Sys.Date(), "\%Y-\%m"),
    ".pdf"),
  output.dir = paste0(dir, "/PfmReports/", format(Sys.Date(), "\%Y-\%m"))
)
}
\arguments{
\item{output.dir}{}
}
\value{

}
\description{
Title
}
