% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_MPOP_report.R
\name{render_MPOP_report}
\alias{render_MPOP_report}
\title{Render MPOP Report}
\usage{
render_MPOP_report(
  rmd.file = "MPOP.Rmd",
  my.dir = getwd(),
  load.data = TRUE,
  dir.output = paste0(my.dir, "/MPOP"),
  file.name = paste0("MPOPReport-", as.character(format(Sys.Date(), "\%Y-\%m")),
    ".pdf"),
  data.period = Sys.Date() \%m-\% months(1),
  ...
)
}
\arguments{
\item{rmd.file}{character}

\item{my.dir}{character}

\item{load.data}{boolean}

\item{dir.output}{character}

\item{file.name}{character}

\item{data.period}{date}

\item{...}{}
}
\value{

}
\description{
This functions prepares and loads the required
datasets and renders the monthly MPOP report
}
\examples{


}
