% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_perf_report.R
\name{render_PERF_report}
\alias{render_PERF_report}
\title{Render Performance Report}
\usage{
render_PERF_report(
  rmd.file = "perf_report_main.Rmd",
  my.dir = getwd(),
  prep.data = TRUE,
  load.data = TRUE,
  dir.output = paste0(my.dir, "/MonthlyPerfReport"),
  file.name = paste0("MonthlyPerformanceReport-", as.character(format(Sys.Date(),
    "\%Y-\%m")), ".pdf"),
  data.period = Sys.Date() \%m-\% months(1),
  ...
)
}
\arguments{
\item{dir.output}{character}

\item{file.name}{character}

\item{...}{foo}

\item{dir}{character}
}
\value{

}
\description{
This function renders the Monthly Performance Report
to be shared with MPC.
}
