% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_pfm_report.R
\name{render_PFM_report}
\alias{render_PFM_report}
\title{Render PFM Report}
\usage{
render_PFM_report(
  my.dir = getwd(),
  tp.list = NULL,
  prep.data = TRUE,
  load.data = TRUE,
  excluded.list = NULL,
  rmd.file = "pfm_report_main.Rmd",
  output.dir = paste0(my.dir, "/PfmReports/", format(Sys.Date(), "\%Y-\%m")),
  data.period = Sys.Date() \%m-\% months(1),
  include.aggregate = FALSE,
  include.consistency = FALSE,
  include.full.ranking = FALSE
)
}
\arguments{
\item{my.dir}{character}

\item{tp.list}{character}

\item{prep.data}{character}

\item{load.data}{character}

\item{excluded.list}{character}

\item{rmd.file}{character}

\item{output.dir}{character}

\item{data.period}{date}
}
\value{

}
\description{
This functions prepares and loads the required
datasets and renders the monthly Portfolio Manager
reports that are sent to Trading Parties.
}
