% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_report}
\alias{render_report}
\title{Render Report}
\usage{
render_report(
  report.type = NULL,
  my.dir = getwd(),
  prep.data = TRUE,
  load.data = TRUE,
  data.period = Sys.Date() \%m-\% months(1),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
A generic function to prepare and load the required
datasets and run either PFM, Performance
or MPOP report
}
\examples{


}
