% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_monthly_pfm_report.R
\name{run_monthly_pfm_report}
\alias{run_monthly_pfm_report}
\title{Run Monthly PFM Report}
\usage{
run_monthly_pfm_report(
  tp.list = NULL,
  prep.data = TRUE,
  load.data = TRUE,
  output.dir = paste0(dir, "/"),
  excluded.list = c("AQUAFLOW-R")
)
}
\arguments{
\item{excluded.list}{}
}
\value{

}
\description{
Run Monthly PFM Report
}
