% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_MOSL.R
\name{scale_colour_MOSL}
\alias{scale_colour_MOSL}
\title{scale_colour_MOSL}
\usage{
scale_colour_MOSL(palette = "all", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{see \code{\link{MOSL_palettes}} for all available colour palettes}

\item{...}{}
}
\value{

}
\description{
This function offers a MOSL colour scale for ggplot2
}
\examples{

}
