% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_assurance_process.R
\name{upload_to_blob_with_meta}
\alias{upload_to_blob_with_meta}
\title{Upload File to Blob with Metadata}
\usage{
upload_to_blob_with_meta(
  processed.template,
  cont,
  container.path,
  org.id,
  assurance.type,
  file.created,
  template.path
)
}
\arguments{
\item{processed.template}{A processed template ready for upload}

\item{cont}{Azure connection for blob storage}

\item{container.path}{Container path for blob storage}

\item{org.id}{Org ID of org site, determined in earlier step}

\item{assurance.type}{Type of assurance, determined in earlier step}

\item{file.created}{Time of file creation, determined in earlier step}

\item{template.path}{Path of the file, determined in earlier step}
}
\description{
Upload File to Blob with Metadata
}
