% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_chart.R
\name{bubble_chart}
\alias{bubble_chart}
\title{Plot Bubble Graph}
\usage{
bubble_chart(
  df = NULL,
  longunread.graph = F,
  vacancy.graph = F,
  by = "Retailer",
  my.dir = getwd(),
  prep.data = F,
  vertical.sep = 0.15
)
}
\arguments{
\item{df}{dataframe}

\item{longunread.graph}{boolean}

\item{vacancy.graph}{boolean}

\item{by}{character, one of c("Retailer", "Wholesaler", "Pair")}

\item{my.dir}{character}

\item{prep.data}{boolean}

\item{vertical.sep}{numerical}
}
\value{
A bubble graph for MPOP reporting
}
\description{
This function creates a bubble graph, based
on either long unread data or vacancy data,
where on the x-axis is the current vacancy
or longunread level and on the y-axis is the
change in the longunread or vacancy level
compared to the level 12 months ago
}
