% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_directory.R
\name{create_directory}
\alias{create_directory}
\title{Create Directory}
\usage{
create_directory(
  my.dir = utils::choose.dir(),
  foldername = "PerformanceMonitoring",
  projectname = "Project"
)
}
\arguments{
\item{my.dir}{character}

\item{foldername}{character}

\item{projectname}{character}
}
\value{

}
\description{
This function creates the required directory for
running the MOSLR performance package.
}
