% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragon_drop.R
\name{dragon_drop}
\alias{dragon_drop}
\title{Drag and drop files from one location to another based on folder name to a specified directory}
\usage{
dragon_drop(
  origin = utils::choose.dir(caption =
    "Select the origin directory that folders or file(s) should be copied from."),
  destination = utils::choose.dir(caption =
    "Select the destination directory that folders or file(s) should be copied to."),
  folder.name = format(Sys.Date(), format = "\%Y-\%m"),
  folder.ext = NULL,
  sub.folder = NULL,
  file.type = NULL,
  folder.list = "destination",
  new.folder = TRUE,
  distribute = FALSE
)
}
\arguments{
\item{origin}{character}

\item{destination}{character}

\item{folder.name}{character}

\item{folder.ext}{character}

\item{sub.folder}{character}

\item{file.type}{factor}

\item{folder.list}{factor}

\item{new.folder}{character}

\item{distribute}{logical}
}
\value{

}
\description{
Drag and drop files from one location to another based on folder name to a specified directory
}
