% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPOP_data_prep.R
\name{MPOP_data_prep}
\alias{MPOP_data_prep}
\title{Prepare MPOP report data}
\usage{
MPOP_data_prep(
  df.longunread = NULL,
  df.vacancy = NULL,
  tp.details = NULL,
  longunread.table = F,
  vacancy.table = F,
  by = "Retailer",
  my.dir = getwd(),
  load.data = F
)
}
\arguments{
\item{df.longunread}{dataframe}

\item{df.vacancy}{dataframe}

\item{tp.details}{dataframe}

\item{longunread.table}{boolean}

\item{vacancy.table}{boolean}

\item{by}{character}

\item{my.dir}{character}

\item{load.data}{boolean}
}
\value{

}
\description{
Prepares data for bubble charts and/or Trading
Party ranking tables in MPOP reporting
}
\examples{



}
