% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps_create_tracker.R
\name{mps_create_tracker}
\alias{mps_create_tracker}
\title{Create Monthly MPS Tracker (Pre Analysis and PfM Commentary)}
\usage{
mps_create_tracker(
  period = Sys.Date() \%m-\% months(1),
  my.dir = getwd(),
  rda.outputs = paste0(my.dir, "/data/rdata"),
  dir.mps.tracking = paste0(my.dir, "/data/tracking/mps"),
  period.only = TRUE,
  save.output = TRUE,
  keep.vars = FALSE,
  iprp.list = NULL,
  filter.category = c("IPRP: on-track", "Performance normal")
)
}
\arguments{
\item{period}{date}

\item{my.dir}{character}

\item{rda.outputs}{character}

\item{dir.mps.tracking}{character}

\item{period.only}{logical}

\item{save.output}{logical}

\item{keep.vars}{logical}

\item{iprp.list}{character}
}
\value{

}
\description{
Compares the performance across MPS for each Trading Party
by MPS and compares IPRP performance versus planned milestones.
This is done prior to analysis and Pfm commentary being added.
}
