% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps_data_prep.R
\name{mps_data_prep}
\alias{mps_data_prep}
\title{MPS Data Preparation}
\usage{
mps_data_prep(
  my.dir = getwd(),
  mps.data,
  mps.thresholds,
  Standards.details,
  csv.outputs = paste0(my.dir, "/data/outputs"),
  rda.outputs = paste0(my.dir, "/data/rdata"),
  save.output = TRUE,
  DataBase = TRUE
)
}
\arguments{
\item{my.dir}{character}

\item{mps.data}{character}

\item{mps.thresholds}{character}

\item{Standards.details}{character}

\item{csv.outputs}{character}

\item{rda.outputs}{character}

\item{save.output}{logical}
}
\value{

}
\description{
Prepares the raw MPS data so it is ready to be processed.
}
