% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps_performance.R
\name{mps_performance}
\alias{mps_performance}
\title{MPS Performance}
\usage{
mps_performance(
  period.lag = 1,
  mps_list = c("MPS 1", "MPS 2", "MPS 3", "MPS 4", "MPS 5", "MPS 7", "MPS 12", "MPS 16",
    "MPS 17", "MPS 18"),
  my.dir = getwd(),
  csv.outputs = paste0(my.dir, "/data/outputs"),
  rda.outputs = paste0(my.dir, "/data/rdata"),
  dir.mps.tracking = paste0(my.dir, "/data/tracking/mps")
)
}
\arguments{
\item{period.lag}{numeric}

\item{mps_list}{character}

\item{my.dir}{character}

\item{csv.outputs}{character}

\item{rda.outputs}{character}

\item{dir.mps.tracking}{character}
}
\value{

}
\description{
Compares the performance across OPS for a given period
for each Trading Party; and compares IPRP performance
versus planned milestones.
}
