% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps_plot_perf_graph.R
\name{mps_plot_perf_graph}
\alias{mps_plot_perf_graph}
\title{Plot Performance Graph (MPS)}
\usage{
mps_plot_perf_graph(
  my.dir = getwd(),
  data = NULL,
  load.data = TRUE,
  melt.data = TRUE,
  include.iprp = FALSE,
  TradParty,
  Standard,
  filter.date.from = "2018-04-01",
  filter.date.to = Sys.Date(),
  action.points = FALSE,
  sec.axis.name = "",
  fill.label = "Task Volume",
  fill.manual = "azure3",
  y.lab = "Tasks",
  x.lab = "",
  graph.title = NULL,
  sub.title = NULL
)
}
\arguments{
\item{my.dir}{character}

\item{data}{data.frame}

\item{load.data}{logical}

\item{melt.data}{logical}

\item{include.iprp}{logical}

\item{TradParty}{character}

\item{Standard}{character}

\item{filter.date.from}{date}

\item{filter.date.to}{date}

\item{action.points}{logical}

\item{sec.axis.name}{character}

\item{fill.label}{character}

\item{fill.manual}{character}

\item{y.lab}{character}

\item{x.lab}{chracter}

\item{graph.title}{character}

\item{sub.title}{character}
}
\value{

}
\description{
This function produces a grasph with two #' independent
axis. The left-handside corresponds to a single variable
plotted as a bar, and the right-handside corresponds to
multiple lines.
}
