% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_data_prep.R
\name{ops_data_prep}
\alias{ops_data_prep}
\title{OPS Data Preparation}
\usage{
ops_data_prep(
  my.dir = getwd(),
  conf.loc = NULL,
  csv.outputs = paste0(my.dir, "/data/outputs"),
  rda.outputs = paste0(my.dir, "/data/rdata"),
  save.output = TRUE,
  DataBase = TRUE
)
}
\arguments{
\item{my.dir}{character}

\item{csv.outputs}{character}

\item{rda.outputs}{character}

\item{save.output}{boolean}

\item{DataBase}{boolean}

\item{ops.data}{character}

\item{ops.details}{character}

\item{ops.thresholds}{character}
}
\value{

}
\description{
Prepares the raw OPS data  so it is ready to be processed.
}
