% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_data_prep.R
\name{ops_data_prep}
\alias{ops_data_prep}
\title{OPS Data Preparation}
\usage{
ops_data_prep(
  my.dir = getwd(),
  ops.data = utils::read.csv(paste0(my.dir, "/data/inputs/OPS_data.csv")),
  ops.thresholds = utils::read.csv(paste0(my.dir, "/data/inputs/ops_thresholds.csv")),
  csv.outputs = paste0(my.dir, "/data/outputs"),
  rda.outputs = paste0(my.dir, "/data/rdata"),
  save.output = TRUE
)
}
\arguments{
\item{my.dir}{character}

\item{ops.data}{character}

\item{csv.outputs}{character}

\item{rda.outputs}{character}

\item{save.output}{logical}
}
\value{

}
\description{
Prepares the raw OPS data  so it is ready to be processed.
}
