% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_data_prep.R
\name{ops_data_prep}
\alias{ops_data_prep}
\title{OPS Data Preparation}
\usage{
ops_data_prep(
  dir = getwd(),
  ops.data = paste0(dir, "/data/inputs/OPS_data.csv"),
  ops.thresholds.kpi = paste0(dir, "/data/inputs/ops_thresholds_kpi.csv"),
  ops.thresholds.api = paste0(dir, "/data/inputs/ops_thresholds_api.csv"),
  tp.details = paste0(dir, "/data/inputs/tp_details.csv"),
  filter.date = "2019-04-01",
  csv.outputs = paste0(dir, "/data/outputs"),
  rda.outputs = paste0(dir, "/data/rdata")
)
}
\arguments{
\item{dir}{character}

\item{ops.data}{character}

\item{ops.thresholds.kpi}{character}

\item{ops.thresholds.api}{character}

\item{tp.details}{character}

\item{filter.date}{character}

\item{csv.outputs}{character}

\item{rda.outputs}{character}
}
\value{

}
\description{
Prepares the raw OPS data  so it is ready to be processed.
}
