% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_perf_graph.R
\name{plot_perf_graph}
\alias{plot_perf_graph}
\title{Plot Performance Graph}
\usage{
plot_perf_graph(
  data,
  include.iprp = TRUE,
  sec.axis.name = "Proportion",
  fill.manual = "azure3",
  y.lab = "Volume of tasks",
  x.lab = "Date",
  tp.name = FALSE,
  graph.title = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{include.iprp}{logical}

\item{sec.axis.name}{character}

\item{fill.manual}{character}

\item{y.lab}{character}

\item{x.lab}{chracter}

\item{tp.name}{logical}

\item{graph.title}{character}
}
\value{

}
\description{
This function produces a grasph with two #' independent
axis. The left-handside corresponds to a single variable
plotted as a bar, and the right-handside corresponds to
multiple lines.
}
