% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_perf_graphs_all.R
\name{plot_perf_graphs_all}
\alias{plot_perf_graphs_all}
\title{Create Performance graphs}
\usage{
plot_perf_graphs_all(
  my.dir = getwd(),
  conf.loc = NULL,
  df.mps = NULL,
  df.ops = NULL,
  load.data = TRUE,
  output.dir = paste0(my.dir, "/graphs"),
  run.parallel = FALSE,
  action.points = FALSE,
  mps.graphs = TRUE,
  ops.graphs = TRUE,
  iprp.graphs = TRUE,
  ops.iprp.graphs = TRUE,
  DataBase = TRUE
)
}
\arguments{
\item{my.dir}{character}

\item{df.mps}{dataframe}

\item{df.ops}{dataframe}

\item{load.data}{boolean}

\item{output.dir}{character}

\item{run.parallel}{boolean}

\item{action.points}{boolean}

\item{mps.graphs}{boolean}

\item{ops.graphs}{boolean}

\item{iprp.graphs}{boolean}

\item{ops.iprp.graphs}{boolean}

\item{DataBase}{boolean}
}
\description{
This function loads the MPS and OPS data
and creates all MPS and OPS performance graphs
}
