% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_PERF_report.R
\name{render_PERF_report}
\alias{render_PERF_report}
\title{Render Performance Report}
\usage{
render_PERF_report(
  rmd.file = "MonthlyPerformanceReport.Rmd",
  dir = getwd(),
  dir.output = paste0(dir, "/MonthlyPerfReport"),
  file.name = paste0("MonthlyPerformanceReport-", as.character(format(Sys.Date(),
    "\%Y-\%m")), ".pdf"),
  ...
)
}
\arguments{
\item{dir}{character}

\item{dir.output}{character}

\item{file.name}{character}

\item{...}{foo}
}
\value{

}
\description{
This function renders the Monthly Performance Report
to be shared with MPC.
}
