% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_MOSL.R
\name{scale_colour_MOSL}
\alias{scale_colour_MOSL}
\title{scale_colour_MOSL}
\usage{
scale_colour_MOSL(
  colours = NULL,
  reverse = FALSE,
  palette = NULL,
  discrete = TRUE,
  ...
)
}
\arguments{
\item{colours}{Specifies the MOSL colours when palette is NULL. When not specified, all colours are selected. See \code{\link{MOSL_colours}} for all available colours}

\item{reverse}{Boolean variable to indicate whether to reverse the order of the scale or not}

\item{palette}{See \code{\link{MOSL_palettes}} for all available colour palettes}

\item{discrete}{Boolean variable to indicate whether to use a discrete scale or not}

\item{...}{}
}
\description{
This function offers a MOSL colour scale for ggplot2
}
