% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_MOSL.R
\name{scale_fill_MOSL}
\alias{scale_fill_MOSL}
\title{scale_fill_MOSL}
\usage{
scale_fill_MOSL(
  colours = NA,
  reverse = FALSE,
  palette = NA,
  discrete = TRUE,
  ...
)
}
\arguments{
\item{colours}{Specifies the MOSL colours when palette is NA. When not specified, all colours are selected. See \code{\link{MOSL_colours}} for all available colours.}

\item{reverse}{Boolean variable to indicate whether to reverse the order of the scale or not}

\item{palette}{see \code{\link{MOSL_palettes}} for all available colour palettes}

\item{discrete}{Boolean variable to indicate whether to use a discrete scale or not}

\item{...}{}
}
\value{

}
\description{
This function offers a MOSL fill scale for ggplot2
}
\examples{

}
